refactor: write resolve function from scratch (wip)
