+++ packages/openapi-parser/tests/diff.test.ts
@@ -37,33 +37,31 @@ const expectedErrors = {
   ],
 }
 
-// const invalidFiles = [
-//   // TODO: max call stack size exceeded or timeout hit
-//   'packages/openapi-parser/tests/files/xerocomxero_accounting.yaml',
-//   'packages/openapi-parser/tests/files/xtrfeu.yaml',
-//   'packages/openapi-parser/tests/files/webflowcom.yaml',
-//   'packages/openapi-parser/tests/files/amazonawscomathena.yaml',
-//   'packages/openapi-parser/tests/files/amazonawscomce.yaml',
-//   'packages/openapi-parser/tests/files/amazonawscomconnect.yaml',
-//   'packages/openapi-parser/tests/files/opentrialslocal.yaml',
-//   'packages/openapi-parser/tests/files/urlboxio.yaml',
-//   'packages/openapi-parser/tests/files/bbccouk.yaml',
-//   'packages/openapi-parser/tests/files/azurecomwindowsiot-windowsiotservices.yaml',
-//   'packages/openapi-parser/tests/files/ote-godaddycomdomains.yaml',
-//   'packages/openapi-parser/tests/files/googleapiscomfirebaserules.yaml',
-//   // TODO: files just failing
-// ]
-
-// const files = (await glob('./packages/openapi-parser/tests/files/*.yaml'))
-//   .filter((file) => !invalidFiles.includes(file))
-//   // Aphabetic
-//   .sort()
-// // Random
-// // .sort(() => Math.random() - 0.5)
-
-const files = [
-  //'packages/openapi-parser/tests/files/bbccouk.yaml'
+const invalidFiles = [
+  // TODO: max call stack size exceeded or timeout hit
+  'packages/openapi-parser/tests/files/xerocomxero_accounting.yaml',
+  'packages/openapi-parser/tests/files/xtrfeu.yaml',
+  'packages/openapi-parser/tests/files/webflowcom.yaml',
+  'packages/openapi-parser/tests/files/amazonawscomathena.yaml',
+  'packages/openapi-parser/tests/files/amazonawscomce.yaml',
+  'packages/openapi-parser/tests/files/amazonawscomconnect.yaml',
+  'packages/openapi-parser/tests/files/opentrialslocal.yaml',
+  'packages/openapi-parser/tests/files/urlboxio.yaml',
+  'packages/openapi-parser/tests/files/bbccouk.yaml',
+  'packages/openapi-parser/tests/files/azurecomwindowsiot-windowsiotservices.yaml',
+  'packages/openapi-parser/tests/files/ote-godaddycomdomains.yaml',
+  'packages/openapi-parser/tests/files/googleapiscomfirebaserules.yaml',
+  // TODO: files just failing
 ]
+
+const files = (await glob('./packages/openapi-parser/tests/files/*.yaml'))
+  .filter((file) => !invalidFiles.includes(file))
+  // Aphabetic
+  .sort()
+// Random
+// .sort(() => Math.random() - 0.5)
+
+// const files = ['packages/openapi-parser/tests/files/bbccouk.yaml']
 
 /**
  * This test suite parses a large number of real-world OpenAPI files
